+++ CHANGELOG.md
@@ -63,6 +63,7 @@ Learn more at [cortexjs.io/math-json/](https://cortexjs.io/math-json/).
   a line before the component would make the component crash. Now the component
   is correctly disconnected, then reconnected and preserve its state across the
   disconnection.
+- **#960** Typing "e^pi" would result in `e\pi` instead of `e^\pi`.
 
 ## 0.67.0 (2021-05-21)